# back/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.QuestionResponse:
    properties:
      answer:
        type: boolean
      created_user_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      selection_1:
        type: string
      selection_2:
        type: string
      selection_3:
        type: string
      selection_4:
        type: string
      statement:
        type: string
      sub_category_id:
        type: integer
      title:
        type: string
      updated_user_id:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://twitter.com/Chuke_yamaha
  description: Api for IchimonItto App
  title: IchimonItto API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: クッキーにログインセッションを記録する。
      parameters:
      - description: Login Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK - Successfully authenticated
        "400":
          description: Bad Request - Invalid request format or parameters
          schema:
            type: string
        "401":
          description: Unauthorized - Invalid login credentials
          schema:
            type: string
      summary: ログイン
      tags:
      - users
...
  /questions/user/{createdUserId}:
    get:
      consumes:
      - application/json
      description: 指定された作成者IDで作成された問題をすべて検索し、結果を返します。
      parameters:
      - description: Created User ID
        in: path
        name: createdUserId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.QuestionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: 作成者IDに基づく問題の検索
      tags:
      - question
  /signup:
    post:
      consumes:
      - application/json
      description: 新しいユーザーを登録します。
      parameters:
      - description: Register User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserResponse'
      summary: ユーザー登録
      tags:
      - users
...
schemes:
- https
swagger: "2.0"